requirements.txt
flask
flask-cors
weasyprint
num2words
python-dateutil